Add routes component and refactoring
