+++ src/components/date/Date.js
@@ -21,7 +21,8 @@ class Clock extends Component {
   render() {
     return (
       <div className="Date">
-        <h2>{this.state.date.toLocaleTimeString()}</h2>
+        <h2>The Time is: <span className="Date-time">{this.state.date.toLocaleTimeString()}</span>
+        </h2>
       </div>
     );
   }
