--- src/components/recentsposts/RecentsPosts.css
@@ -1,16 +1,0 @@
-.RecentsPosts {
-  text-align: left;
-}
-.RecentsPosts h1 {
-  letter-spacing: 0.2rem;
-  background-color: #151a45;
-  color: #fff;
-  padding: 1rem 2rem;
-  margin-bottom: 1.5rem;
-}
-ul {
-  list-style-type: none;
-}
-li {
-  font-size: 1.6rem;
-}
