--- src/components/recentsposts/RecentsPosts.js
@@ -1,20 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import "./RecentsPosts.css";
-
-const RecentsPosts = ({ posts }) => {
-  return (
-    <div className='RecentsPosts'>
-      <h1>RECENTS POSTS</h1>
-      <ul>
-        {posts.map(post => (
-          <Link to={`/posts/${post.title.toLowerCase()}`}>
-            <li key={post.id}>{post.title}</li>
-          </Link>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-export default RecentsPosts;
